Move versioned-cache unit tests
